requirements.txt
django>=4.2
djangorestframework>=3.14
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.30
langchain-ollama>=0.1.0
ollama>=0.1.0
chromadb>=0.4.22
requests>=2.31
pydantic>=2.7
django-cors-headers>=4.3.1
